move appliance police to a own pytest plugin
